+++ .dev/dev-allinone.js
@@ -2,7 +2,7 @@
 * All In One Generator
 *
 * Description: Creates "kioskboard-aio.js" file automatically from "kioskboard.js" and "kioskboard.css" files.
-* Version: 1.0.0
+* Version: 2.0.0
 * Author: Furkan MT ('https://github.com/furcan')
 * Copyright 2021 All In One Generator, MIT Licence ('https://opensource.org/licenses/MIT')
 */
@@ -14,7 +14,7 @@ const CleanCSS = require('clean-css');
 const Constants = require('./dev-constants');
 
 // Constants
-const thisFilePath = '.dev/helpers/dev-allinone.js';
+const thisFilePath = '.dev/dev-allinone.js';
 
 // Get File content as text: begin
 const getFileContentAsTextByType = (filePath, fileType, clean) => {