+++ CHANGELOG.md
@@ -1,3 +1,23 @@
+@2.0.0
+* **Removed:** The `KioskBoard.Merge()` method has been removed. (This method already has been deprecated in v1.4.0)
+
+* **Changed:** `KioskBoard.Init()` function name has been changed to `KioskBoard.init()`.
+
+* **Changed:** `KioskBoard.Run()` function name has been changed to `KioskBoard.run()`.
+
+* **Changed:** Auto-generated `kioskboard-aio.js` file has been moved from `src/all-in-one` folder to `build` folder.
+
+* **Changed:** The `specialCharactersObject` option has been changed to `keysSpecialCharsArrayOfStrings`. An Array of Strings can be set to override the built-in special characters. e.g. => `["#", "$", "%", "+", "-", "*"]`
+
+* **Fixed:** Custom key with multiple characters: ([#31](https://github.com/furcan/KioskBoard/issues/31))
+
+* **Added:** The `keysNumpadArrayOfNumbers` option has been added: An Array of Numbers can be set to override the built-in numpad keys. (From 0 to 9, in any order.) e.g. => `[1, 2, 3, 4, 5, 6, 7, 8, 9, 0]` - ([#30](https://github.com/furcan/KioskBoard/issues/30))
+
+
+-----
+-----
+-----
+
 @1.4.0
 * **Fixed:** The dispatcher issue on the input change event has been fixed: ([#11](https://github.com/furcan/KioskBoard/issues/11))
 * **Fixed:** The current text selection issue has been fixed: ([#19](https://github.com/furcan/KioskBoard/issues/19))
@@ -60,16 +80,16 @@
 -----
 
 @1.0.0
-* A pure JavaScript library for using the Virtual Keyboard is ready to use.
+* KioskBoard is ready to use.
 
 -----
 
 @1.0.0-beta.02
-* A pure JavaScript library for using the Virtual Keyboard. First Release.
+* KioskBoard. First Release.
 
 -----
 
 @1.0.0-beta.01
-* A pure JavaScript library for using the Virtual Keyboard. Coming Soon.
+* KioskBoard. Coming Soon.
 
 -----